Move proofs interfaces into proofs package
