+++ proofs/interface.go
@@ -1,4 +1,8 @@
-package lightclient
+package proofs
+
+import (
+	lc "github.com/tendermint/light-client"
+)
 
 // Prover is anything that can provide proofs.
 // Such as a AppProver (for merkle proofs of app state)
@@ -18,7 +22,7 @@ type Prover interface {
 type Proof interface {
 	BlockHeight() uint64
 	// Validates this Proof matches the checkpoint
-	Validate(Checkpoint) error
+	Validate(lc.Checkpoint) error
 	// Marshal prepares for storage
 	Marshal() ([]byte, error)
 	// Data extracts the query result we want to see
